DSND-2242: Refactor config and test config/utils
