+++ src/main/java/uk/gov/companieshouse/filinghistory/consumer/delta/DeltaConfig.java
@@ -1,17 +1,16 @@
-package uk.gov.companieshouse.filinghistory.consumer.config;
+package uk.gov.companieshouse.filinghistory.consumer.delta;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import java.util.function.Supplier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import uk.gov.companieshouse.api.InternalApiClient;
-import uk.gov.companieshouse.api.http.ApiKeyHttpClient;
+import uk.gov.companieshouse.sdk.manager.ApiSdkManager;
 
 @Configuration
-public class AppConfig {
+public class DeltaConfig {
 
     @Bean
     public ObjectMapper objectMapper() {
@@ -21,13 +20,7 @@ public class AppConfig {
     }
 
     @Bean
-    Supplier<InternalApiClient> internalApiClientSupplier(
-            @Value("${api.api-key}") String apiKey,
-            @Value("${api.api-local-url}") String apiUrl) {
-        return () -> {
-            InternalApiClient internalApiClient = new InternalApiClient(new ApiKeyHttpClient(apiKey));
-            internalApiClient.setBasePath(apiUrl);
-            return internalApiClient;
-        };
+    Supplier<InternalApiClient> internalApiClientSupplier() {
+        return ApiSdkManager::getPrivateSDK;
     }
 }
