+++ src/main/resources/application.properties
@@ -5,9 +5,6 @@ consumer.max-attempts=${MAX_ATTEMPTS:5}
 consumer.backoff-delay=${BACKOFF_DELAY:100}
 consumer.concurrency=${CONCURRENT_LISTENER_INSTANCES:10}
 
-api.api-local-url=${API_LOCAL_URL:localhost:9090}
-api.api-key=${CHS_API_KEY:apiKey}
-
 management.endpoints.enabled-by-default=false
 management.endpoints.web.base-path=/
 management.endpoints.web.path-mapping.health=healthcheck
