+++ src/test/java/uk/gov/companieshouse/filinghistory/consumer/kafka/KafkaUtils.java
@@ -3,17 +3,17 @@ package uk.gov.companieshouse.filinghistory.consumer.kafka;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 
-public final class TestUtils {
+final class KafkaUtils {
 
     static final String MAIN_TOPIC = "filing-history-delta";
     static final String RETRY_TOPIC = "filing-history-delta-filing-history-delta-consumer-retry";
     static final String ERROR_TOPIC = "filing-history-delta-filing-history-delta-consumer-error";
     static final String INVALID_TOPIC = "filing-history-delta-filing-history-delta-consumer-invalid";
 
-    private TestUtils() {
+    private KafkaUtils() {
     }
 
-    public static int noOfRecordsForTopic(ConsumerRecords<?, ?> records, String topic) {
+    static int noOfRecordsForTopic(ConsumerRecords<?, ?> records, String topic) {
         int count = 0;
         for (ConsumerRecord<?, ?> ignored : records.records(topic)) {
             count++;
